Reorganize sql to gcs operators.
