+++ tests/contrib/operators/test_mssql_to_gcs_operator.py
@@ -63,7 +63,7 @@ class MsSqlToGoogleCloudStorageOperatorTest(unittest.TestCase):
         self.assertEqual(op.filename, JSON_FILENAME)
 
     @mock.patch('airflow.contrib.operators.mssql_to_gcs.MsSqlHook')
-    @mock.patch('airflow.contrib.operators.mssql_to_gcs.GoogleCloudStorageHook')
+    @mock.patch('airflow.contrib.operators.sql_to_gcs.GoogleCloudStorageHook')
     def test_exec_success_json(self, gcs_hook_mock_class, mssql_hook_mock_class):
         """Test successful run of execute function for JSON"""
         op = MsSqlToGoogleCloudStorageOperator(
@@ -95,7 +95,7 @@ class MsSqlToGoogleCloudStorageOperatorTest(unittest.TestCase):
         mssql_hook_mock.get_conn().cursor().execute.assert_called_once_with(SQL)
 
     @mock.patch('airflow.contrib.operators.mssql_to_gcs.MsSqlHook')
-    @mock.patch('airflow.contrib.operators.mssql_to_gcs.GoogleCloudStorageHook')
+    @mock.patch('airflow.contrib.operators.sql_to_gcs.GoogleCloudStorageHook')
     def test_file_splitting(self, gcs_hook_mock_class, mssql_hook_mock_class):
         """Test that ndjson is split by approx_max_file_size_bytes param."""
         mssql_hook_mock = mssql_hook_mock_class.return_value
@@ -126,7 +126,7 @@ class MsSqlToGoogleCloudStorageOperatorTest(unittest.TestCase):
         op.execute(None)
 
     @mock.patch('airflow.contrib.operators.mssql_to_gcs.MsSqlHook')
-    @mock.patch('airflow.contrib.operators.mssql_to_gcs.GoogleCloudStorageHook')
+    @mock.patch('airflow.contrib.operators.sql_to_gcs.GoogleCloudStorageHook')
     def test_schema_file(self, gcs_hook_mock_class, mssql_hook_mock_class):
         """Test writing schema files."""
         mssql_hook_mock = mssql_hook_mock_class.return_value
@@ -135,7 +135,7 @@ class MsSqlToGoogleCloudStorageOperatorTest(unittest.TestCase):
 
         gcs_hook_mock = gcs_hook_mock_class.return_value
 
-        def _assert_upload(unused_bucket, obj, tmp_filename, unused_mime_type=None, unused_gzip=False):
+        def _assert_upload(bucket, obj, tmp_filename, mime_type, gzip):  # pylint: disable=unused-argument
             if obj == SCHEMA_FILENAME:
                 with open(tmp_filename, 'rb') as file:
                     self.assertEqual(b''.join(SCHEMA_JSON), file.read())
